Refactor frame decoding and color correction
