+++ src/highdicom/frame.py
@@ -4,23 +4,25 @@ from io import BytesIO
 import numpy as np
 from PIL import Image
 from pydicom.dataset import Dataset
+from pydicom.encaps import encapsulate
 from pydicom.pixel_data_handlers.numpy_handler import pack_bits
 from pydicom.pixel_data_handlers.rle_handler import rle_encode_frame
 from pydicom.uid import (
     ExplicitVRLittleEndian,
     ImplicitVRLittleEndian,
-    JPEGBaseline,
     JPEG2000Lossless,
+    JPEGBaseline,
     RLELossless,
 )
+
 
 logger = logging.getLogger(__name__)
 
 
 def encode_frame(
-        array: np.ndarray,
-        metadata: Dataset,
-    ) -> bytes:
+    array: np.ndarray,
+    metadata: Dataset,
+) -> bytes:
     """Encodes pixel data of an individual frame.
 
     Parameters
@@ -40,11 +42,26 @@ def encode_frame(
 
     Raises
     ------
+    AttributeError
+        When required attribute is missing in `metadata`.
     ValueError
         When transfer syntax is not supported.
 
     """
     transfer_syntax_uid = metadata.file_meta.TransferSyntaxUID
+    required_attributes = {
+        'Rows',
+        'Columns',
+        'BitsAllocated',
+        'BitsStored',
+        'SamplesPerPixel',
+    }
+    for attr in required_attributes:
+        if not hasattr(metadata, attr):
+            raise AttributeError(
+                'Cannot encode frame. '
+                f'Image metadata is missing required attribute "{attr}".'
+            )
     bits_allocated = metadata.BitsAllocated
     rows = metadata.Rows
     cols = metadata.Columns
@@ -109,3 +126,101 @@ def encode_frame(
                 f'Transfer Syntax "{transfer_syntax_uid}" is not supported.'
             )
     return data
+
+
+def decode_frame(
+    value: bytes,
+    metadata: Dataset,
+) -> np.ndarray:
+    """Decodes pixel data of an individual frame.
+
+    Parameters
+    ----------
+    value: bytes
+        Pixel data of a frame (potentially compressed in case
+        of encapsulated format encoding, depending on the transfer syntax)
+    metadata: pydicom.dataset.Dataset
+        Metadata of the corresponding image dataset
+
+    Returns
+    -------
+    numpy.ndarray
+        Decoded pixel data
+
+    Raises
+    ------
+    AttributeError
+        When required attribute is missing in `metadata`.
+    ValueError
+        When transfer syntax is not supported.
+
+    """
+    # The pydicom library does currently not support reading individual frames.
+    # This hack creates a small dataset containing only a single frame, which
+    # can then be decoded using the pydicom API.
+    ds = Dataset()
+    ds.file_meta = metadata.file_meta
+    required_attributes = {
+        'Rows',
+        'Columns',
+        'BitsAllocated',
+        'BitsStored',
+        'SamplesPerPixel',
+        'PhotometricInterpretation',
+        'PixelRepresentation',
+    }
+    for attr in required_attributes:
+        try:
+            setattr(ds, attr, getattr(metadata, attr))
+        except AttributeError:
+            raise AttributeError(
+                'Cannot decode frame. '
+                f'Image metadata is missing required attribute "{attr}".'
+            )
+
+    if metadata.SamplesPerPixel > 1:
+        attr = 'PlanarConfiguration'
+        try:
+            setattr(ds, attr, getattr(metadata, attr))
+        except AttributeError:
+            raise AttributeError(
+                'Cannot decode frame. '
+                f'Image metadata is missing required attribute "{attr}".'
+            )
+
+    transfer_syntax_uid = metadata.file_meta.TransferSyntaxUID
+    rows = metadata.Rows
+    columns = metadata.Columns
+    samples_per_pixel = metadata.SamplesPerPixel
+    bits_allocated = metadata.BitsAllocated
+
+    if transfer_syntax_uid.is_encapsulated:
+        if (transfer_syntax_uid == JPEGBaseline and
+                metadata.PhotometricInterpretation == 'RGB'):
+            # RGB color images, which were not transformed into YCbCr color
+            # space upon JPEG compression, need to be handled separately.
+            # Pillow assumes that images were transformed into YCbCr color
+            # space prior to JPEG compression. However, with photometric
+            # interpretation RGB, no color transformation was performed.
+            # Setting the value of "mode" to YCbCr signals Pillow to not
+            # apply any color transformation upon decompression.
+            n_pixels = rows * columns * samples_per_pixel
+            n_bytes = n_pixels * bits_allocated // 8
+            compressed_pixels = value[:n_bytes]
+            image = Image.open(BytesIO(compressed_pixels))
+            color_mode = 'YCbCr'
+            image.tile = [(
+                'jpeg',
+                image.tile[0][1],
+                image.tile[0][2],
+                (color_mode, ''),
+            )]
+            image.mode = color_mode
+            image.rawmode = color_mode
+            return np.asarray(image)
+        else:
+            ds.PixelData = encapsulate(frames=[value])
+    else:
+        ds.PixelData = value
+
+    return ds.pixel_array
